requirements/base.txt
PyYAML==6.0.3
Jinja2==3.1.6
click==8.3.1
requests==2.32.5
semver==3.0.4
marshmallow==4.2.2
packaging==26.0
tenacity==9.1.4
